feat: new logging apiroute and upgrade logger
This commit introduces a new route (src/app/api/logs/route.ts) to get the last 100 log lines as a JSON response. Additionally, the level of logging in winston has been upgraded from 'error' to 'info'. The irrelevant logger error message from the main page has been removed and the log files are added to the .gitignore file for convenience.